Fixes, cleanup and refacto
